+++ kopf/testing.py
@@ -3,7 +3,7 @@ Helper tools to test the Kopf-based operators.
 
 This module is a part of the framework's public interface.
 """
-from kopf._kits.runner import KopfRunner
+from kopf._kits.runners import KopfRunner
 
 __all__ = [
     'KopfRunner',
